Made a banner for README.md
Hey bro, I made a banner for the benin db, I tried to make it good, but the readme doesn't really reflect any personality of the repo so I just went with my own imagination which is not very good. But I tried it to fit well with the rust theme, like orange and red.